use u_int8_t and u_int16_t instead of size_t for some quiz related fields

+++ include/kanji_tools/quiz/GroupQuiz.h
@@ -16,7 +16,7 @@ public:
   // - All: include all members (as long as they have readings)
   enum MemberType { Jouyou, JLPT, Frequency, All };
 
-  GroupQuiz(const QuizLauncher&, size_t question, bool showMeanings,
+  GroupQuiz(const QuizLauncher&, u_int16_t question, bool showMeanings,
             const GroupData::List&, MemberType);
 private:
   // 'GroupEntryWidth' is the width required for 'qualified name', 'pinyin' and
